requirements.txt
discord.py>=2.3.0
PyNaCl>=1.5.0  # Required for Voice
asyncio
# pyttsx3 removed (requires heavy system libs, causes crashes on minimal hosts)